Default display driver is now mac,x11
Now wine versions that support Mac Driver will launch using it, only falling back to X11 when needed.
Wineskin.app screen options "mac" radio button" is now disabled if engine does not support it.

Now Mac/X11 setting are mac,x11 for priority to mac driver and x11 fallback, and x11,mac for the reverse, this is useful if a port is made using x11 but the end user does not have XQuartz installed it will fallback to using mac driver if the engine supports it.